Added Repair action for machine break game

+++ skills-sellers/Entities/Actions/ActionReparer.cs
@@ -2,5 +2,5 @@ namespace skills_sellers.Entities.Actions;
 
 public class ActionReparer : Action
 {
-    public int? RepairChances { get; set; }
+    public double? RepairChances { get; set; }
 }
